[APU/FreeRTOS+Linux] Update UZP framework (adapter card identification) based on Tobi's input

+++ vitis/software/FreeRTOS/uz/uz_PLATFORM/uz_platform.h
@@ -54,7 +54,7 @@ uint32_t uz_platform_macread(uint8_t eeprom, uint8_t *addr);
 uint32_t uz_platform_macread_primary(uint8_t *addrbuf_p);
 
 #if (UZ_PLATFORM_CARDID==1)
- uint32_t uz_platform_cardread(uint8_t slot, uz_platform_eeprom_group000models_t* model_p, int* revision_p, int* serial_p);
+ uint32_t uz_platform_cardread(uint8_t slot, uz_platform_eeprom_group000models_t* model_p, uint8_t* revision_p, uint16_t* serial_p);
 #endif
 
 #endif
